(2) restructure web framework source struct .
