--- Abstract/Version.go
@@ -1,5 +1,0 @@
-package Abstract
-
-const (
-	Version = "v1.4.9.4 pre-release"
-)
