+++ Test/render_test.go
@@ -5,8 +5,8 @@ import (
 	"encoding/xml"
 	"errors"
 	"github.com/golang/protobuf/proto"
-	"github.com/maxzhang1985/yoyogo/ActionResult"
 	"github.com/maxzhang1985/yoyogo/Test/testdata/protoexample"
+	"github.com/maxzhang1985/yoyogo/WebFramework/ActionResult"
 	"github.com/stretchr/testify/assert"
 	"github.com/ugorji/go/codec"
 	"html/template"
