+++ Version.go
@@ -1,0 +1,5 @@
+package main
+
+const (
+	Version = "v1.4.9.5 pre-release"
+)
